### 2018-08-17 9217c51
Reimplemented ReinsertionDistance to use O(n lg n) alg for longest common subsequence
