### 2018-08-17 de272cf
reimplemented Kendall Tau distance in O(n lg n) time by counting inversions with modified mergesort
