### 2018-09-01 131c36a
added distance metrics on sequences (String, arrays, etc)
