+++ tests/org/cicirello/sequences/PrimitiveValueTests.java
@@ -1,0 +1,164 @@
+/*
+ * Copyright 2018 Vincent A. Cicirello, <https://www.cicirello.org/>.
+ *
+ * This file is part of JavaPermutationTools (https://jpt.cicirello.org/).
+ *
+ * JavaPermutationTools is free software: you can 
+ * redistribute it and/or modify it under the terms of the GNU 
+ * General Public License as published by the Free Software 
+ * Foundation, either version 3 of the License, or (at your 
+ * option) any later version.
+ *
+ * JavaPermutationTools is distributed in the hope 
+ * that it will be useful, but WITHOUT ANY WARRANTY; without even 
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A 
+ * PARTICULAR PURPOSE.  See the GNU General Public License for more 
+ * details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with JavaPermutationTools.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+package org.cicirello.sequences;
+
+import org.junit.*;
+import static org.junit.Assert.*;
+
+/**
+ * JUnit 4 tests for the PrimitiveValue class.
+ */
+public class PrimitiveValueTests {
+	
+	private static final double DELTA_0 = 0.0; // for comparisons that should be exact
+	private static final float DELTA_F_0 = 0.0f; // for comparisons that should be exact
+	
+	@Test
+	public void testPrimitiveValueLong() {
+		long value = Integer.MAX_VALUE;
+		value = value + 100L;
+		PrimitiveValue pv = new PrimitiveValue(value);
+		assertEquals("to long", value, pv.longValue());
+		assertEquals("to int", (int)value, pv.intValue());
+		assertEquals("to short", (short)value, pv.shortValue());
+		assertEquals("to byte", (byte)value, pv.byteValue());
+		assertEquals("to char", (char)value, pv.charValue());
+		assertEquals("to double", (double)value, pv.doubleValue(), DELTA_0);
+		assertEquals("to float", (float)value, pv.floatValue(), DELTA_F_0);
+		assertTrue("to boolean true", pv.booleanValue());
+		assertFalse("to boolean false", (new PrimitiveValue(0L)).booleanValue());
+	}
+	
+	@Test
+	public void testPrimitiveValueInt() {
+		int value = Integer.MAX_VALUE;
+		value = value - 100;
+		PrimitiveValue pv = new PrimitiveValue(value);
+		assertEquals("to long", (long)value, pv.longValue());
+		assertEquals("to int", value, pv.intValue());
+		assertEquals("to short", (short)value, pv.shortValue());
+		assertEquals("to byte", (byte)value, pv.byteValue());
+		assertEquals("to char", (char)value, pv.charValue());
+		assertEquals("to double", (double)value, pv.doubleValue(), DELTA_0);
+		assertEquals("to float", (float)value, pv.floatValue(), DELTA_F_0);
+		assertTrue("to boolean true", pv.booleanValue());
+		assertFalse("to boolean false", (new PrimitiveValue(0)).booleanValue());
+	}
+	
+	@Test
+	public void testPrimitiveValueShort() {
+		short value = (short)1024;
+		PrimitiveValue pv = new PrimitiveValue(value);
+		assertEquals("to long", (long)value, pv.longValue());
+		assertEquals("to int", (int)value, pv.intValue());
+		assertEquals("to short", value, pv.shortValue());
+		assertEquals("to byte", (byte)value, pv.byteValue());
+		assertEquals("to char", (char)value, pv.charValue());
[…]
