### 2018-09-01 0fcab34
updated JOSS draft
