### 2018-09-01 #887 @cicirello
@yochannah  Thanks.  I'll take a look at the list of reviewers. Also, since submitting I actually added a package of distance metrics on sequences to the library, and just added a short paragraph toward the end of the paper to indicate such.  Anyway, the version number when I submitted JOSS paper was 1.0, but current version number is now 1.1 (in case that matters for the review).
