### 2018-09-10 #887 @ChrisJefferson
I am happy to review this package. I know Java quite well, but am a little rusty with building / classpaths / junit.

@cicirello  Could you please add explicit instructions for building the package, running the tests and running the examples?
