### 2018-09-10 #887 @cicirello
@ChrisJefferson Yes, I'll add instructions for building, running the tests, running the examples.  I'm out of town at the moment, but will be back by tomorrow night, so I'll add this on Wednesday morning.  Thank you for reviewing.
