### 2018-09-11 #887 @yochannah
@ChrisJefferson Thank you so much for volunteering to review! I'm hunting for a second reviewer (feel free to suggest if you know anyone), and once we find them we can kick off the full review. In the meantime feel free to look at the package and browse the review guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
