+++ README.md
@@ -23,6 +23,19 @@ Compiled jar files of the library are in the /lib folder.
 The /examples folder includes source code to illustrate how to use the library.  The /replication
 folder includes source code that recreates results found in publications associated with the library.
 
+## Building the Library
+
+The /build directory contains an ant build file.  The build file also executes the JUnit tests, and the build directory contains
+the relevant jar files for the JUnit libraries.  If you prefer, you can replace these with the latest versions available
+from https://junit.org/junit4/ (just be sure to edit the property fields in the build.xml to point to the locations of the JUnit jar files).
+
+To execute the build process do one of the following.  If your working directory is the build directory, then simply execute
+`ant` from the command line.  If your working directory is the parent of build, then execute: `ant -f build/build.xml`
+
+The default of the provided ant build file, compiles all source files and all JUnit tests, executes all test cases,
+and generates the jar file of the library. The build process will terminate on any test case failures.
+
+
 ## Contribute
 
 Report bugs, suggestions, feature requests, etc via the [issues tracker](https://github.com/cicirello/JavaPermutationTools/issues).  If you would like to directly contribute new code, you may also fork the repository, and create pull requests.
