### 2018-09-12 e3e7ff6
updated readme.md for building library
