### 2018-09-12 1aa9bde
example folder readme updated with explicit running instructions (also to use markdown)
