+++ examples/readme.txt
@@ -1,15 +1,0 @@
-Examples Directory Readme:
-
-The source for examples is contained in this directory, and
-currently includes the following examples:
-
-org.cicirello.examples.jpt.AverageDistance
-org.cicirello.examples.jpt.TableOfDistances
-
-Read the comments in the source for a description of what they
-demonstrate, etc.
-
-Make sure the jar file of the library is in your classpath when
-you compile and/or run any of these. Specifically, make sure jpt1.jar is in your
-classpath (or any later version--I intend to ensure later versions
-are backwards compatible).
