### 2018-09-12 6e09795
fixed missing bullets
