### 2018-09-12 b4b3eab
updated readme for replication folder to use markdown and to include explicit instructions for running
