### 2018-09-12 #887 @cicirello
@ChrisJefferson I added a subsection to the repository's readme on building the library, as well as a directory (build) with an ant build file.  Assuming compilation is successful, the build will also execute all of the JUnit tests, and terminate if any tests fail (you'll see any relevant test results during the build, such as number of test cases, etc).  If you don't usually use ant, it can be obtained here: https://ant.apache.org/

For running the example programs, and the programs that replicate experimental data from published papers, I have updated the readme files in the examples and replication directories to include explicit instructions (including setting the classpath).
