### 2018-09-16 #887 @motib
I have agreed to review the submission although I haven't written Java for many years.....

I successfully installed ant and built the package. When I run the test case I get an error message:

D:\joss>java -classpath d:\joss\javaperm\lib;d:\joss\javaperm\testbin org.cicirello.permutations.PermutationTestCases
Error: Main method not found in class org.cicirello.permutations.PermutationTest
Cases, please define the main method as:
   public static void main(String[] args)
or a JavaFX application class must extend javafx.application.Application

@cicirello Could you please help?!
