### 2018-09-16 #950 @whedon
[REVIEW]: JavaPermutationTools: A Java Library of Permutation Distance Metrics
**Submitting author:** @cicirello (<a href="http://orcid.org/0000-0003-1072-8559">Vincent A Cicirello</a>)
**Repository:** <a href="https://github.com/cicirello/JavaPermutationTools" target ="_blank">https://github.com/cicirello/JavaPermutationTools</a>
**Version:** 1.0
