### 2018-09-16 #950 @cicirello
@motib Just replying to your question from the pre-review issue.... JUnit tests don't have main methods. They're executed differently (either in an IDE or with a runner, such as the one within ant).   In this case, I have the ant build script set up to run all of the JUnit tests during the build.

When you run ant to build the project, you'll see the JUnit test results among the output. The relevant lines are labeled as JUnit. After each JUnit test class is executed, ant will output a line indicating the number of test cases, some other summary data, and indicating if any failed and why.

The ant build script is also configured to terminate if any tests fail. So you'll only get a successful build if they all pass.

There are also example programs in the examples and replication folders, which you can run from command line (see the readmes in those folders).
