### 2018-09-17 #950 @motib
Thanks, @cicirello, I eventually got it to work.

I suggest that the readme explain how to compile and run the programs using only "javac" and "java" in directories not part of the distribution. I imagine that some researchers in algorithms are inexperienced programmers who are not familiar with the entire ecosystem (e.g. ant), and who might need help with nested packages. But you know your community well and if you think that I am misjudging their level of software expertise, feel free to leave everything as it.
