### 2018-09-17 #950 @motib
After reviewing according to the review criteria, I recommend "accept with minor revisions". Suggestions follow:

JOSS paper

The paper is too long (the author guidelines state 250-1000 words). If you still think that a long paper is necessary, please give _subsections_ according to the JOSS guidelines: * A summary describing the high-level functionality, and purpose of the software for a diverse, non-specialist audience, * A clear statement of need that illustrates the purpose of the software, * Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it

Given the guideline of "non-specialist audience", in addition to the research papers, there should be a reference to a textbook presentation or tutorial on algorithm permutations. (My 1974! edition of Aho, Hopcroft and Ullman only has one exercise on permutations. Cormen, Leiserson and Rivest might have more.)

Example usage

OK, but consider the suggestion I I wrote in the previous post.

API documentation

The link in the local copy of "index.html" doesn't work. I could only access it through the link on "https://jpt.cicirello.org/".

The API itself is excellent, but the guidelines recommend "including example inputs and outputs". This doesn't need to be (shouldn't be?) in the API, but I suggest that more example programs be included in the project.
