### 2018-09-17 0fc0a55
Added alternative build script (Windows batch file)
