### 2018-09-17 90e8847
added subsection to readme on running example programs
