### 2018-09-17 215eb99
Edited paper based on reviewer suggestions
