+++ paper/paper.bib
@@ -91,6 +91,31 @@ number = {1/2},
 pages = {81--93}
 }
 
+@book{knuth1,
+author = {Donald E. Knuth},
+title = {The Art of Computer Programming, Volume 1, Fundamental Algorithms},
+edition = {3rd},
+year = {1997},
+publisher = {Addison Wesley}
+}
+
+@book{knuth2,
+author = {Donald E. Knuth},
+title = {The Art of Computer Programming, Volume 2, Seminumerical Algorithms},
+edition = {3rd},
+year = {1998},
+publisher = {Addison Wesley}
+}
+
+@book{knuth3,
+author = {Donald E. Knuth},
+title = {The Art of Computer Programming, Volume 3, Sorting and Searching},
+edition = {2nd},
+year = {1998},
+publisher = {Addison Wesley}
+}
+
+
 @article{lee58, 
 author={C. Lee}, 
 journal={IRE Transactions on Information Theory}, 
