+++ paper/paper.md
@@ -13,7 +13,7 @@ authors:
 affiliations:
   - name: Computer Science, School of Business, Stockton University, Galloway, NJ 08205
     index: 1
-date: 20 August 2018
+date: 8 August 2018
 bibliography: paper.bib
 ---
 
@@ -27,55 +27,69 @@ between sets (e.g., instructors to courses at a fixed time).  There are applicat
 where measuring the distance between a pair of permutations is necessary.  For example,
 a recommender system may want to compare the similarity of two individuals' preferences 
 for music, perhaps to make song recommendations.  Depending upon the application, 
-the permutation features most important to distance calculation may be one or more of 
+the permutation features most important to distance calculation may be 
 the absolute positions of the elements (e.g., for one-to-one mappings), the adjacency 
-of elements (e.g., the earlier routing example), or the general precedence of pairs of 
+of elements (e.g., the routing example), or the general precedence of pairs of 
 elements (e.g., music preference example).  Thus, it should be no surprise that there
 exists a large variety of permutation distance metrics in the research literature.
+Knuth's seminal books on algorithms [@knuth1; @knuth2; @knuth3] include permutation related 
+algorithms employed by our library such as mixed radix representation, permutation 
+inverse computation, etc.  
 
-The motivation and origin of this library of permutation distance metrics is our research 
-on evolutionary computation, and other metaheuristic approaches, for solving permutation 
-optimization problems.  A permutation optimization problem is a problem where solutions 
-are represented by permutations of the elements of some set, and where the objective is to 
-maximize or minimize some function.  For example, a solution to a traveling salesperson 
-problem (TSP) is a permutation of the set of cities, and the objective is to find the 
-permutation that corresponds to the tour of the cities of minimal cost.
+The motivation and origin of this library is our research on fitness landscape analysis 
+for permutation optimization problems [@cicirello2016; @cicirello2014; @cicirello2013].  A 
+permutation optimization problem is a problem where solutions are represented by 
+permutations of the elements of some set, and where the objective is to maximize 
+or minimize some function.  For example, a solution 
+to a traveling salesperson problem (TSP) is a permutation of the set of cities, and the 
+objective is to find the permutation that corresponds to the tour of the cities of minimal cost.
+During the course of our research, we developed a Java library of permutation distance 
+metrics.  Most of the distance metrics in the literature are described mathematically with 
+no source code available.  Thus, our library offers convenient access to efficient 
+implementations of a variety of distance metrics that share a common programmatic 
+interface.  The library also provides distance metrics on sequences (strings and arrays of 
+various types); where unlike a permutation, a sequence may contain multiple copies of the same element. 
 
-There are at least two applications of permutation distance that come from evolutionary 
-computation.  The first application is to maintain population diversity.  The various forms 
-of evolutionary computation, such as genetic algorithms, genetic programming, etc, solve 
-problems through simulated evolutionary processes [@mitchell].  They maintain a population of solutions 
-to the problem at hand, and this population evolves over many generations using operators 
-such as mutation, crossover, etc.  Just as it is in natural evolution, a diverse gene 
-pool is important to evolutionary computation.  In later generations, if variation among 
-the members of the population declines, then the evolutionary search can stagnate.  Approaches 
-to population management [@sevaux2005], such as in scatter search [@campos2005; @marti2005], 
-attempt to maintain diversity among the population of solutions.  They require a means of 
-measuring distance between population members.
+The source repository (https://github.com/cicirello/JavaPermutationTools) 
+contains source code of the library, programs that provide example 
[…]
