### 2018-09-17 #950 @cicirello
@motib I've made changes and/or added things based on your comments.  I believe I've accounted for almost all of your suggestions.  Here's a summary of what I've done, and what I haven't gotten to yet.

First, list of what I've done so far based on your suggestions:

(1) In the build directory, in addition to the ant build.xml, there is now a Windows batch script that uses only javac (to compile), jar to generate the jar of the library, and java to execute the JUnit tests.  If someone has the java jdk installed, they will already have all of these.

(2) I edited the sections of the readme for the repository to document how to build.  Now two sections, one for ant, and one for this alternative process.

(3) I added a section to the readme for the repository explaining how to run the example programs (although I do still direct them to the readme in the example folders for more details).

(4) I fixed the bug in the index.html for the docs, where the link to the API docs was broken if viewed locally (I had a leading / in the link, which was fine when viewed via the web, but not when viewed locally).

(5) I added a sentence in the paper, referring to Knuth's book series for general coverage of permutation related algorithms.  The CLRS algorithms book you suggested actually doesn't have much on permutations.

(6) I reorganized and edited the text of the paper in various ways.  e.g., Now multiple sections, including one on statement of need.

(7) In editing the text of the paper, I also cut down length slightly.  It is now just under the upper end of the suggested length range at 998 words (including title, author, keywords, etc,).

Next, what I haven't addressed (yet)....  I think the only one of your suggestions that I haven't addressed yet is adding more examples.  Are there any parts of the library that you think would especially benefit from more examples?  Anyway, I'll probably add more examples on Wednesday (I teach most of tomorrow).
