### 2018-09-18 #950 @motib
@cicirello Really nice now.

I suggest one example for each of the five top-level packages. You could probably do this with almost no work by adapting the JUnit tests programs to be standalone programs in the examples directory.

No rush. Tomorrow is a holiday here, so I won't get to it again before Thursday.
