### 2018-09-18 fae0e33
minor edits
