+++ paper/paper.bib
@@ -18,6 +18,24 @@ pages = {75--83},
 publisher = {ACM}
 } 
 
+@techreport{cicirello2018a,
+author = {Vincent A. Cicirello},
+title = {Classification of Permutation Distance Metrics for Fitness Landscape Analysis},
+year = {2018},
+month = {October},
+institution = {Stockton University},
+type = {preprint}
+}
+
+@techreport{cicirello2018b,
+author = {Vincent A. Cicirello},
+title = {Kendall Tau Sequence Distance: Extending Kendall Tau from Ranks to Sequences},
+year = {2018},
+month = {October},
+institution = {Stockton University},
+type = {preprint}
+}
+
 @article{cicirello2016,
 author = {Vincent A. Cicirello},
 title = {The Permutation in a Haystack Problem and the Calculus of Search Landscapes},
