### 2018-09-19 e41f3d2
added several example programs illustrating usage of library
