+++ examples/org/cicirello/examples/jpt/PermutationExamples.java
@@ -1,0 +1,115 @@
+/*
+ * Copyright 2018 Vincent A. Cicirello, <https://www.cicirello.org/>.
+ *
+ * This file is part of JavaPermutationTools (https://jpt.cicirello.org/).
+ *
+ * JavaPermutationTools is free software: you can 
+ * redistribute it and/or modify it under the terms of the GNU 
+ * General Public License as published by the Free Software 
+ * Foundation, either version 3 of the License, or (at your 
+ * option) any later version.
+ *
+ * JavaPermutationTools is distributed in the hope 
+ * that it will be useful, but WITHOUT ANY WARRANTY; without even 
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A 
+ * PARTICULAR PURPOSE.  See the GNU General Public License for more 
+ * details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with JavaPermutationTools.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package org.cicirello.examples.jpt;
+
+import org.cicirello.permutations.*;
+import java.util.HashMap;
+
+/**
+ * <p>Basic examples of the use of the classes from the 
+ * org.cicirello.permutations package, including the
+ * {@link org.cicirello.permutations.Permutation Permutation} and 
+ * {@link org.cicirello.permutations.PermutationIterator PermutationIterator} classes.</p>
+ *
+ *
+ * @author Vincent A. Cicirello, https://www.cicirello.org/
+ */
+public class PermutationExamples {
+	
+	public static void main(String[] args) {
+		
+		System.out.println("If you are running this example program, make sure you also read the source code and comments.");
+		System.out.println("Output of example program in isolation of the source and comments won't tell you much.");
+		
+		// You can construct Permutation objects in several ways.
+		// Here are a couple examples (see API for additional constructors).
+		
+		// Generating a random permutation of length 10:
+		Permutation p1 = new Permutation(10);
+		
+		// Initializing a Permutation from an array of ints.  Note: array must contain unique elements
+		// and must be permutation of the integers from 0 to N-1 (where N is length).
+		int[] a = { 3, 0, 4, 2, 1, 5};
+		Permutation p2 = new Permutation(a);
+		
+		// Making a copy of a permutation:
+		Permutation p3 = new Permutation(p1);
+		
+		// Constructing a Permutation from an int in mixed-radix representation.
+		int length = 5;
+		int permID = 10;
+		Permutation p4 = new Permutation(length, permID);
+		
+		// The Permutation class overrides Object.toString so you can easily generate text output that 
+		// includes descriptions of permutations.  For example:
+		System.out.println("Permutation p1 is: " + p1);
+		
+		// The Permutation class overrides Object.hashCode so you can use Permutation objects as the
+		// key to a HashMap, HashSet, etc.
+		HashMap<Permutation,String> map = new HashMap<Permutation,String>();
+		map.put(p1, "A permutation maps to me.");
+		
+		// Object.equals overridden to check if 2 permutation objects consist in the same permutation of integers.
+		if (p2.equals(p3)) {
+			System.out.println("p2 and p3 are the same.");
+		} else {
+			System.out.println("p2 and p3 are not the same.");
+		}
+		
+		// length method returns the length, get method accesses a specific element.
+		System.out.println("i\tp1.get(i)");
+		for (int i = 0; i < p1.length(); i++) {
+			System.out.println(i + "\t" + p1.get(i));
+		}
+		
+		// getInverse returns an array of ints that is the inverse of the Permutation
+		int[] inv = p1.getInverse();
+		Permutation pInv = new Permutation(inv);
[…]
