+++ examples/readme.md
@@ -5,6 +5,10 @@ currently includes the following examples:
 
 * org.cicirello.examples.jpt.AverageDistance
 * org.cicirello.examples.jpt.TableOfDistances
+* org.cicirello.examples.jpt.PermutationExamples
+* org.cicirello.examples.jpt.SimpleDistanceExamples
+* org.cicirello.examples.jpt.SequenceExamples
+* org.cicirello.examples.jpt.SequenceDistanceExamples
 
 Read the comments in the source code itself for a description of what they
 demonstrate, etc.
@@ -28,4 +32,8 @@ of the repository, then you can run the examples from the command line with the 
 ```
 java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.AverageDistance
 java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.TableOfDistances
+java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.PermutationExamples
+java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.SimpleDistanceExamples
+java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.SequenceExamples
+java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.SequenceDistanceExamples
 ```
