+++ README.md
@@ -65,6 +65,10 @@ of the repository, then you can run these programs from the command line with th
 ```
 java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.AverageDistance
 java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.TableOfDistances
+java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.PermutationExamples
+java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.SimpleDistanceExamples
+java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.SequenceExamples
+java -cp "lib/jpt1.jar;exbin" org.cicirello.examples.jpt.SequenceDistanceExamples
 java -cp "lib/jpt1.jar;exbin" org.cicirello.replication.flairs2013.Flairs2013
 java -cp "lib/jpt1.jar;exbin" org.cicirello.replication.ieeetevc2016.FDC
 ```

[…]
