### 2018-09-19 868bd3d
streamlined text of paper in several places
