+++ paper/paper.md
@@ -19,35 +19,35 @@ bibliography: paper.bib
 
 # Summary
 
-Permutations are used to represent a wide variety of ordered data.  For example,
+Permutations can represent a wide variety of ordered data.  For example,
 a permutation may represent an individual's preferences (or ranking) of a collection
 of products such as books or music.  Or perhaps a permutation may represent a route
 for delivering a set of packages.  Permutations can also represent one-to-one mappings 
 between sets (e.g., instructors to courses at a fixed time).  There are applications 
 where measuring the distance between a pair of permutations is necessary.  For example,
-a recommender system may want to compare the similarity of two individuals' preferences 
-for music, perhaps to make song recommendations.  Depending upon the application, 
+a recommender system may assess the similarity of two individuals' preferences 
+for music to make song recommendations.  Depending upon the application, 
 the permutation features most important to distance calculation may be 
-the absolute positions of the elements (e.g., for one-to-one mappings), the adjacency 
-of elements (e.g., the routing example), or the general precedence of pairs of 
-elements (e.g., music preference example).  Thus, it should be no surprise that there
-exists a large variety of permutation distance metrics in the research literature.
-Knuth's seminal books on algorithms [@knuth1; @knuth2; @knuth3] include permutation related 
-algorithms employed by our library such as mixed radix representation, permutation 
+the absolute positions of the elements (e.g., one-to-one mappings), the adjacency 
+of elements (e.g., the routing example), or general precedence of pairs of 
+elements (e.g., music preferences).  Thus, it is no surprise that there
+are many permutation metrics in the research literature.
+Knuth's seminal books on algorithms [@knuth1; @knuth2; @knuth3] cover permutation related 
+algorithms more generally such as mixed radix representation, permutation 
 inverse computation, etc.  
 
 The motivation and origin of this library is our research on fitness landscape analysis 
-for permutation optimization problems [@cicirello2018a; @cicirello2016; @cicirello2014; @cicirello2013].  A 
-permutation optimization problem is a problem where solutions are represented by 
-permutations of the elements of some set, and where the objective is to maximize 
+for permutation optimization problems [@cicirello2018a; @cicirello2016; @cicirello2014; @cicirello2013].  In a 
+permutation optimization problem, solutions are represented by 
+permutations of some set, and the objective is to maximize 
 or minimize some function.  For example, a solution 
-to a traveling salesperson problem (TSP) is a permutation of the set of cities, and the 
-objective is to find the permutation that corresponds to the tour of the cities of minimal cost.
-During the course of our research, we developed a Java library of permutation distance 
+to a traveling salesperson problem (TSP) is the permutation of the set of cities
+that corresponds to the minimal cost tour of the cities.
+During our research, we developed a Java library of permutation distance 
 metrics.  Most of the distance metrics in the literature are described mathematically with 
 no source code available.  Thus, our library offers convenient access to efficient 
-implementations of a variety of distance metrics that share a common programmatic 
-interface.  The library also provides distance metrics on sequences (strings and arrays of 
+implementations of a variety of metrics with a common programmatic 
+interface.  The library also provides metrics on sequences (strings and arrays of 
 various types); where unlike a permutation, a sequence may contain multiple copies of the same element. 
 
 The source repository (https://github.com/cicirello/JavaPermutationTools) 
@@ -68,21 +68,21 @@ Evolutionary computation, such as genetic algorithms, solve problems through
 simulated evolutionary processes [@mitchell].  They maintain a population of solutions 
 to the problem, and this population evolves over many generations using operators 
[…]
