### 2018-09-19 11e474b
column widths adjusted in 2nd table
