### 2018-09-19 973850d
fiddled with column widths of tables
