### 2018-09-19 07b9fbc
minor edits
