### 2018-09-19 207e2a7
minor edits and more table column width fiddling
