### 2018-09-19 #950 @cicirello
@motib I added example programs for each of the top level packages, as you suggested.  I also trimmed down the text of the paper a bit more (and made the tables look nicer).
