### 2018-09-20 #950 @motib
Thanks, @cicirello.

I ran the programs with only one problem: the ieeetevc2016.FDC program ran out of memory (heap space). I added "-Xms1g" to the "java" command and it ran, although it took a long time. (I vaguely remembered that such a parameter exists, although it doesn't appear on the "java" usage, so I had to google it....).

Please add to the readme's and program comments an indication of how much memory and time is needed. (At work I am stuck with a 32bit 3Ghz 4GB computer, but, again, others might also have primitive computers!!)
