### 2018-09-20 291f39d
warning on memory usage of 1 replication program
org.cicirello.replication.ieeetevc2016.FDC may require increasing heap allocation to JVM in order to run.
