### 2018-09-20 #950 @cicirello
@motib Thanks for catching the potential memory issue with that one experiment replication program.  That one is very memory intensive.  No problems on our server with the default heap allocation, but default is probably higher on a server JVM.

Anyway, I added a paragraph to both the repository's readme, and the readme in the replication folder on this (also mentioning it runs long), indicating to use the -Xms1g parameter if an out of memory error is thrown (and explicitly showed the full command line statement).
