### 2018-09-20 #950 @ChrisJefferson
Someinitial comments:

Most methods don't have checks for invalid inputs -- personally I would have some kind of checked/debugging mode which checked that (for example) permutations are valid. At least one method I think this is more serious -- Permutation::toInteger can easily overflow (particularly as it calculates n!), but the result could still be used to build a new, invalid permutation.

The xor-swap trick, used in various places in Permutation, is both less clear, and slower, than just using a temporary int variable.

It is unclear if you intend to support permutations on 0 points. The tests don't seem to check them, but the code doesn't forbid them.

There are not many tests for many of the distances, checking a permutation against itself, or the result on identity / 1 or 0 point permutations.
