### 2018-09-21 91b21d5
error checking on toInteger, and removed set method
Permutation::toInteger now throws an exception when n > 12 (similar
change to constructor of ReversalDistance).

Removed Permutation::set and changed ReversalDistance::distance to
eliminate use of Permtation::set (the set method had potential to
produce invalid permutation, and there are alternate ways of
implementing things that previously used it).
