### 2018-09-21 23ffb61
Added 0-length permutation tests
Added tests for 0-length permutations for the constructors and methods
of Permutation.  Also edited test methods for the permutation distance
classes to include 0-length permutations among tests.  Fixed bug in 2
distance classes related to 0-length permutations.
