### 2018-09-21 d2aa494
updated documentation of Permutation class
Documented potential bounds exception for methods of Permutation with
index parameters.
