### 2018-09-21 c4de167
eliminated xor trick
