### 2018-09-21 1f9a8c6
added 0-length tests for sequence distances
