+++ src/org/cicirello/permutations/distance/AcyclicEdgeDistance.java
@@ -68,7 +68,5 @@ public class AcyclicEdgeDistance extends AbstractPermutationDistanceMeasurer {
 		}
 		return countNonSharedEdges;
 	}
-	
-
-
 }
+
