### 2018-09-21 1b29429
refactored sequence distances
Moved utility classes used only by the sequence distances into the
org.cicirello.sequences.distance package, and made them package-private.
