+++ src/org/cicirello/sequences/package-info.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2018 Vincent A. Cicirello, <https://www.cicirello.org/>.
- *
- * This file is part of JavaPermutationTools (https://jpt.cicirello.org/).
- *
- * JavaPermutationTools is free software: you can 
- * redistribute it and/or modify it under the terms of the GNU 
- * General Public License as published by the Free Software 
- * Foundation, either version 3 of the License, or (at your 
- * option) any later version.
- *
- * JavaPermutationTools is distributed in the hope 
- * that it will be useful, but WITHOUT ANY WARRANTY; without even 
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A 
- * PARTICULAR PURPOSE.  See the GNU General Public License for more 
- * details.
- *
- * You should have received a copy of the GNU General Public License
- * along with JavaPermutationTools.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
- 
- /**
- * Collection of classes related to manipulating general sequences of various forms (e.g., arrays or various types, Strings, Lists). 
- * @version 2.18.8.20
- * @since 1.1
- */
-package org.cicirello.sequences;
