### 2018-09-21 5b9a373
refactored sequence distances
Moved utility classes for sequence distance into the same package as the
distances, and made those utility classes package-private.
