### 2018-09-21 48c602e
refactored sequence distances
Refactoring of sequence distances made one example program obsolete (was
giving examples of usage of a coupld classes that are now
package-private).
