+++ examples/org/cicirello/examples/jpt/SequenceExamples.java
@@ -1,161 +1,0 @@
-/*
- * Copyright 2018 Vincent A. Cicirello, <https://www.cicirello.org/>.
- *
- * This file is part of JavaPermutationTools (https://jpt.cicirello.org/).
- *
- * JavaPermutationTools is free software: you can 
- * redistribute it and/or modify it under the terms of the GNU 
- * General Public License as published by the Free Software 
- * Foundation, either version 3 of the License, or (at your 
- * option) any later version.
- *
- * JavaPermutationTools is distributed in the hope 
- * that it will be useful, but WITHOUT ANY WARRANTY; without even 
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A 
- * PARTICULAR PURPOSE.  See the GNU General Public License for more 
- * details.
- *
- * You should have received a copy of the GNU General Public License
- * along with JavaPermutationTools.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package org.cicirello.examples.jpt;
-
-import org.cicirello.sequences.*;
-
-/**
- * Basic examples of the use of the classes from the 
- * org.cicirello.sequences package.  Important Note:
- * The classes of the org.cicirello.sequences package 
- * are mainly intended to serve as utility classes for
- * implementing algorithms that operate on sequences,
- * specifically to enable implementing such algorithms
- * on sequences independent of what type of elements
- * are contained in the sequence.
- *
- * @author Vincent A. Cicirello, https://www.cicirello.org/
- */
-public class SequenceExamples {
-	public static void main(String[] args) {
-		
-		System.out.println("If you are running this example program, make");
-		System.out.println("sure you also read the source code and comments.");
-		System.out.println("Output of example program in isolation of the");
-		System.out.println("source and comments won't tell you much.");
-		System.out.println();
-		
-		// The PrimitiveValue class is a wrapper class that supports all 8 of Java's primitive types.
-		// It has been implemented to support specific methods of the PrimitiveSequence class (the get and set methods).
-		// The PrimitiveValue class has simpler functionality than the Java API's Number classes (Integer, Double, etc)
-		// such as no auto-boxing.
-		// The PrimitiveValue class's functionality is limited to wrapping a primitive value of any of the 8 primitive types,
-		// and converting among them.
-		// Some examples:
-		PrimitiveValue v1 = new PrimitiveValue(5);
-		PrimitiveValue v2 = new PrimitiveValue(true);
-		PrimitiveValue v3 = new PrimitiveValue(10.21);
-		int i1 = v1.intValue();
-		boolean b1 = v1.booleanValue();
-		
-		// The org.cicirello.sequences also defines a Sequence interface, and two classes PrimitiveSequence and
-		// ObjectSequence that both implement the Sequence interface.
-		// The purpose of these classes (and interface) is to enable implementing algorithms that operate on
-		// sequences of any type of element without the need to know what type of element is contained in the sequence.
-		
-		// IMPORTANT NOTE: You can use the sequence distance measures of the org.cicirello.sequences.distance package
-		// without directly using the classes and interface of the org.cicirello.sequences package, for which we now
-		// provide examples.  
-		
-		// Some examples of constructing Sequences:
-
-		
-		// sequence of int values
-		int[] iArray = {2, 6, 1, 4, 8, 4, 4, 4, 2};
-		PrimitiveSequence s1 = new PrimitiveSequence(iArray);
-		// sequence of double values
-		double[] dArray = {1.1, 4.0, 3.1, 10.0, -2.1};
-		PrimitiveSequence s2 = new PrimitiveSequence(dArray);
-		
[…]
