### 2018-09-21 37a9389
Updated readme and webpage index file
A refactoring resulted in removal of a package.  Updated list of
packages to reflect this.
