+++ README.md
@@ -9,7 +9,6 @@ https://www.cicirello.org/
 JPT is organized into the following packages:
 * __org.cicirello.permutations__: Permutation class and a class for iterating over permutations.
 * __org.cicirello.permutations.distance__: Collection of permutation distance metric implementations.
-* __org.cicirello.sequences__: Classes and interfaces for representing and manipulating sequences (i.e., Strings, arrays, etc).
 * __org.cicirello.sequences.distance__: Collection of distance metrics on Strings and other sequences.
 * __org.cicirello.math.stats__: Implementations of statistical concepts.
 
