### 2018-09-21 #950 @cicirello
@ChrisJefferson Thank you for your very useful feedback.  Here's a list of what I've done to address your comments.  I think I've covered all of your comments here.

1) Permutation::toInteger modified to throw an exception if permutation length greater than 12 (to avoid int overflow) and comments edited to document this.  Similar change made to ReversalDistance which has a similar limit.

2) On potential for invalid inputs to the various methods of Permutation with index parameters.... Rather than checking, I've now edited documentation to document potential for ArrayIndexOutOfBoundsExceptions.

3) Related to the above, the JUnit tests for the Permutation class tests that the constructors generate valid Permutations.  All of the methods of the Permutation class that change the permutation (except 1 of them), if implemented correctly guarantee result is a valid Permutation.  The one method that does not have this guarantee, I have now removed (the set method).  I was only using set in one other class (ReversalDistance) where there was an alternative way to do the same thing.  So have decided not worth the potential risk of misuse of set.

4) Eliminated the xor-swap trick.

5) I hadn't explicitly thought about 0-length permutations.  I edited test cases (both tests of Permutation class, as well as the tests of the distance metrics) to test with 0-length permutations.  There was one distance metric that didn't handle that case correctly, so fixed that too (everything else handled the 0-length permutation correctly).

6) On distance tests.... All distance metrics already had tests for pairs of identical permutations (see the identicalPermutations method in that test class).  However, I added to it, testing with two references to the same permutation object.  Also, already testing for 1-length permutations (see the identicalPermutations method in that test class).  There is only 1 case of a permutation of length 1: { 0 }, which is a distance of 0 from itself.  Anyway, I have edited the tests to also now test the distance between the 0 length permutation from itself (and fixed 2 distance metrics that threw exceptions on the 0 length case).  Thanks for catching that I neglected 0-length permutations.

7) This one isn't based on your comments: I had a couple classes in a package org.cicirello.sequences that are really just utility classes that serve the purpose of easing the implementation of the distance metrics in org.cicirello.sequences.distance, and which otherwise wouldn't apply outside of that context.    So I moved those utility classes into the org.cicirello.sequences.distance package, and changed access to package-private.
