### 2018-10-08 #950 @yochannah
Hey @ChrisJefferson! Quick status check - have you had any chance to check the paper? 😄
