### 2018-10-15 #950 @ChrisJefferson
Sorry, I got distracted.

I am very happy with the whole library now, except for one issue (although it's a big one in some sense).

Personally, I don't like functions like the `(Permutation p, int length)` permutation constructor, which can produce invalid permutations if given incorrect input data -- I find it's too easy to accidentally make a slip up, and end up with an invalid permutation. In my code I would tend to have not have such functions (or at the least, label them somehow, with `unsafe` in their name), to make users think hard before using them.

Now, I wouldn't want to reject out of hand on that basis, as I realise the trade-off between speed and safety is a personal one (as I've got older, I've found no amount of speed is worth the occasions I have spent literal weeks tracking down a bug in a massive software project).

Therefore my only request is that either these constructors are made safe (possibly an `Assert`, or some other system which can be disabled if you really don't want to pay the cost of checking the permutation is valid although I don't know if they are trendy/acceptable in Java nowdays?), or alternatively to clearly document that some constructors can produce invalid permutations if given incorrect input data (this is technically already there, but could be made ALL CAPS, so no-one misses it).
