### 2018-10-15 2f84258
Previously unsafe constructor now safe.
Redefined the partial copy constructor's behavior such that it always
produces a valid permutation.
