### 2018-10-15 #950 @cicirello
@ChrisJefferson The constructor with parameters `Permutation(Permutation p, int length)` will now only generate valid permutations.  I redefined its behavior (see the updated documentation of this constructor) slightly, but just enough that it could be implemented such that any parameter values can be handled.  Also replaced the old test cases for this constructor to test the new behavior.

Related to this, I updated the documentation of the constructor with parameters `Permutation(int[] p)` to document the exception that is thrown if it is passed an array that contains duplicates or element values outside the interval [0,n).  It was already doing that validation check, but the exception wasn't documented in the javadoc comments.  It now is.

So now there are no constructors or methods that produce invalid permutations, and parameter values that would lead to such are checked with exceptions thrown and documented with `@throws`.
