### 2018-10-15 #950 @ChrisJefferson
Excellent, thanks.

I am now completely happy with this package.

This is not at all a requirement, but I notice when looking in github ( for example https://github.com/cicirello/JavaPermutationTools/blob/master/src/org/cicirello/permutations/Permutation.java#L92 ), that the formatting looks weird. Mixed tabs+spaces? Or a tab size different to github's?
