+++ src/org/cicirello/permutations/Permutation.java
@@ -44,39 +44,39 @@ public final class Permutation implements Serializable, Iterable<Permutation>
 	private static final long serialVersionUID = 1L;
 	
 	/**
-     * Initializes a random permutation of n integers.  Uses
+	 * Initializes a random permutation of n integers.  Uses
 	 * java.util.concurrent.ThreadLocalRandom as the source of efficient random number generation.
-     * @param n the length of the permutation
-     */
+	 * @param n the length of the permutation
+	 */
 	public Permutation(int n) {
 		this(n, ThreadLocalRandom.current());
 	}
  
-    /**
-     * Initializes a random permutation of n integers.
-     * @param n the length of the permutation
+	/**
+	 * Initializes a random permutation of n integers.
+	 * @param n the length of the permutation
 	 * @param r A source of randomness.
-     */
-    public Permutation(int n, SplittableRandom r) {
-        permutation = new int[n];
-        for (int i = 0; i < n; i++) {
-            permutation[i] = i;   
-        }
-        scramble(r);
-    }
+	 */
+	public Permutation(int n, SplittableRandom r) {
+		permutation = new int[n];
+		for (int i = 0; i < n; i++) {
+			permutation[i] = i;   
+		}
+		scramble(r);
+	}
 	
 	/**
-     * Initializes a random permutation of n integers.
-     * @param n the length of the permutation
+	 * Initializes a random permutation of n integers.
+	 * @param n the length of the permutation
 	 * @param r A source of randomness.
-     */
-    public Permutation(int n, Random r) {
-        permutation = new int[n];
-        for (int i = 0; i < n; i++) {
-            permutation[i] = i;   
-        }
-        scramble(r);
-    }
+	 */
+	public Permutation(int n, Random r) {
+		permutation = new int[n];
+		for (int i = 0; i < n; i++) {
+			permutation[i] = i;   
+		}
+		scramble(r);
+	}
 	
 	/**
 	* Initializes a specific permutation from an integer in mixed radix form representing the chosen
@@ -89,8 +89,8 @@ public final class Permutation implements Serializable, Iterable<Permutation>
 	public Permutation(int n, int value) {
 		permutation = new int[n];
 		for (int i = 0; i < n; i++) {
-            permutation[i] = i;   
-        }
+			permutation[i] = i;   
+		}
 		for (int i = 0; i < n-1; i++) {
 			int j = i + value % (n-i);
 			int temp = permutation[j];
@@ -104,7 +104,7 @@ public final class Permutation implements Serializable, Iterable<Permutation>
 	
 	/**
 	 * Initializes a permutation of n integers to be identical to the elements of an array.
-     * @param p An array of integers. Each of the integers in the interval [0, p.length) must occur exactly one time each.
+	 * @param p An array of integers. Each of the integers in the interval [0, p.length) must occur exactly one time each.
 	 * @throws IllegalArgumentException if p either contains duplicates, or contains any negative elements, 
 	 *         or contains any elements equal or greater than p.length.
 	 */
@@ -118,39 +118,39 @@ public final class Permutation implements Serializable, Iterable<Permutation>
 		permutation = p.clone();
 	}
 	
-    /**
-     * Initializes a permutation of n integers to be identical to a given
-     * permutation.
-     * @param p the given permutation.
-     */
-    public Permutation(Permutation p) {
-        permutation = new int[p.permutation.length];
-        System.arraycopy(p.permutation, 0, permutation, 0, p.permutation.length);
-    }
-    
-    /**
[…]
