### 2018-10-21 #950 @yochannah
Okay - @ChrisJefferson If you're happy with the package can you run through the checklist at the top of this issue and make sure you're happy to check them all? Thanks! Once that's done we can move on to publishing this
