### 2018-10-24 #950 @ChrisJefferson
My only issue is that there have been some commits, as part of the review, which should be part of a new release.
