### 2018-10-24 #950 @cicirello
@ChrisJefferson @yochannah I just created a new release to include all commits made during the review process.
