### 2018-11-06 #950 @arfon
@cicirello - your paper is now accepted into JOSS :zap: :rocket: :boom:
