### 2018-11-06 #950 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
